+++ notes-app/notes-app.js
@@ -1,4 +1,3 @@
-// const { Console } = require("console");
 
 const notes = getSavedNotes();
 
@@ -17,24 +16,6 @@ document.querySelector('#create-note').addEventListener('click', function(e){
     renderNotes(notes, filters);
 })
 
-// render the notes data on the browser
-const renderNotes = function (notes, filters) {
-    const filteredNotes = notes.filter(function (note) {
-        return note.title.toLowerCase().includes(filters.searchText.toLowerCase());
-    })
-
-    document.querySelector('#notes-con').innerHTML = ''
-
-    filteredNotes.forEach(function (note) {
-        const noteEl = document.createElement('p');
-        if(note.title.length > 0){
-            noteEl.textContent = note.title;
-        }else{
-            noteEl.textContent = 'unamed note';
-        }
-        document.querySelector('#notes-con').appendChild(noteEl);
-    })
-}
 renderNotes(notes, filters);
 
 // target search input
